extend RouterServiceFactory trait
